backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.5.0
slowapi>=0.1.9
python-dotenv>=1.0.0
sqlmodel>=0.0.14
sqlalchemy>=2.0.23
google-generativeai>=0.3.0
